requirements.txt
python-telegram-bot>=20.0
aiohttp>=3.8.0
psutil>=5.9.0
python-dotenv>=0.19.0
logging>=0.4.9.6